Same field formatting for summarise fields as filter fields
